requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-google-genai
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
pandas
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
